Create OpenBuildings2FeatureClass.py
This script is to convert the CSV file(s) of Google Open Buildings to a Feature Class. This script uses ArcPy modules, so your environment should have ArcPy library. Simply, load the script to your Tool Box on your ArcGIS Pro project and run it. The target Open Buildings CSV file must be stored in the folder where the script is located.